Correct FEN loading (+ change files' extension)
