+++ .gitignore
@@ -1,3 +1,4 @@
 Nelly
 Nelly.exe
 *.log
+*.out
